requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
requests>=2.31.0,<3.0.0
google-genai>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytesseract>=0.3.10,<1.0.0
json_repair>=0.29.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
Pillow>=10.0.0,<12.0.0
pypdf>=4.0.0,<6.0.0
pdf2image>=1.16.0,<2.0.0
